Made changes to the way the importing/loading of configs were handled.
Added a new directory: "../cstrike/addons/eventscripts/gungame51/scripts/configs" which include both a "custom" and "included" folder to store cfglib addons for configs to be loaded.

Added a new command in core/cfg/__init__.py called getConfigList() that returns a list of configs located in the above-listed directories. If no arguments are provided, it returns all config names. Other arguments are "included", "custom", and "main" ("main" is the list of primary GunGame configs).

Added __init__.py to the custom and included folder in "../gungame51/scripts/configs" for handling the "from ... import *"