WOW! DependencyErrors were being raised due to the order of config execution and loading/unloading of addon dependencies and conflicts. Therefore, this commit is primarily to fix these errors.
I take charge of executing the configs, and created a global variable to toggle between "True" and "False" (the variable is called "cfgExecuting"). I used gamethread.delayed(0) to toggle this to true to give the config enough time to execute. In the self.server_cvar() function, I check to see if the config is executing.

Now, the cool part is, if the config IS executing, I check the default value set forth by our cfglib.AddonCFG instance. If this has been deviated from, we raise an error. However, if the value is the same, and there would be a DependencyError, we do not allow the error to be raised.

Furthermore, if a config is not being executed, a DependencyError will be raised as the result of an admin/user setting the value to something that it should not be.

I also added es.set(name, 0|1) to the DependencyErrors, so that the variable would automatically be toggled back to its original value, instead of having to do so in server_cvar.

I removed config.execute() from all configs located in core/cfg/files since I now handle it in the core/cfg/__init__.py.

Also, I changed the directory structure:
  "../gungame51/scripts/configs"
to:
  "../gungame51/scripts/cfg".