MEGA-COMMIT TIME!!!
All files in this commit contain the same basic change. If a class is designed to only contain one instance, I have coded in a "def __new__" method into the class so that only one instance of that class will ever be returned instead of creating a new instance like standard classes do.

This will help to limit having to import the class instance as opposed to the class. This is not quite Python "standard", however, this is coded for GunGame, and it limits the possibility of someone creating a class instance that has nothing to with the currently-running GunGame.

I replaced all imports of class instance variables for the "singleton" classes with the import of the class, and changed all references to the class instance variables to use the class instead:

"__configs__.somemethod()" is now "ConfigManager().somemethod()".

I have tested the crap out of this, and I hope we don't need to revert any changes. I haven't found anything wrong, but feel free to let me know should you find any problems. This commit has the potential to break GunGame5.1 entirely.

Take special not that the class PriorityAddon() is one of the classes that I changed. Now, you simply import PriorityAddon and do "PriorityAddon().append('gg_script_name')".