./scripts/included/gg_final_fight/gg_final_fight.py: * Changed "gg_deadstrip" to "gg_dead_strip".
ALL files ending in "_config":
* Added ".addFlag('notify')" to CVARs needing to fire the server_cvar event.


NOTE TO DEVS:
Add the notify flag in your config generator ONLY to CVARs that you need to fire the server_cvar event. Please note that all CVARs that are used to load and unload scripts need to have the "notify" flag.

This fix is in response to:
http://code.google.com/p/gungame51/issues/detail?id=22